Split ssr-server.js into multiple files (#659)
* Remove unused export.

Implementation is incomplete, and the `getAssetUrl` that is actually used
comes from `pwa-kit-react-sdk`, not `pwa-kit-runtime`.

* Split ssr-server utils into multiple files.

Smaller files are easier to debug.

* Update jest coverage threshold to reflect new file structure.
